pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tui-meshcore"
version = "0.1.0"
description = "A TUI client for meshcore LoRa mesh networks"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "textual>=0.47.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
radio = [
    "pymc_core",
]

[project.scripts]
tui-meshcore = "tui_meshcore.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tui_meshcore"]
